+++ app/build.gradle.kts
@@ -203,6 +203,7 @@ android {
 
 dependencies {
     // Mockito and Mockito-Kotlin for Kotlin-specific extensions
+
     testImplementation ("org.mockito.kotlin:mockito-kotlin:4.0.0")
     implementation("com.googlecode.libphonenumber:libphonenumber:8.13.37")
     implementation(libs.core.ktx)
