+++ app/src/main/java/com/example/task1/ui/dashboard/DashboardViewModelFactory.kt
@@ -20,7 +20,7 @@ class DashboardViewModelFactory(
 
 }
 //When your ViewModel requires parameters that aren't directly available through default constructors or need to be injected (e.g., a Repository, a context, or any other dependencies), the ViewModelFactory pattern comes into play.
-//Your DashboardViewModel needs an Application instance for its initialization.
+//my DashboardViewModel needs an Application instance for its initialization.
 //This factory takes the Application instance and provides it to the DashboardViewModel when it is created.
 //Use a ViewModelFactory to pass parameters.
 //If your ViewModel has a no-arg constructor, you can create it without a factory:
